Move BlsToExecutionChangeGenerator to spec.testFixtures
